+++ game/db.py
@@ -129,19 +129,6 @@ from dcs.ships import (
     Tanker_Elnya_160,
     ship_map,
 )
-from dcs.task import (
-    AWACS,
-    CAP,
-    CAS,
-    CargoTransportation,
-    Embarking,
-    MainTask,
-    Nothing,
-    PinpointStrike,
-    Reconnaissance,
-    Refueling,
-    Transport,
-)
 from dcs.terrain.terrain import Airport
 from dcs.unit import Ship, Unit, Vehicle
 from dcs.unitgroup import ShipGroup, StaticGroup
@@ -150,7 +137,6 @@ from dcs.vehicles import (
     AirDefence,
     Armor,
     Artillery,
-    Carriage,
     Infantry,
     Unarmed,
     vehicle_map,
@@ -164,6 +150,7 @@ from game.factions.faction_loader import FactionLoader
 from pydcs_extensions.a4ec.a4ec import A_4E_C
 from pydcs_extensions.f22a.f22a import F_22A
 from pydcs_extensions.hercules.hercules import Hercules
+from pydcs_extensions.jas39.jas39 import JAS39Gripen, JAS39Gripen_AG
 from pydcs_extensions.mb339.mb339 import MB_339PAN
 from pydcs_extensions.su57.su57 import Su_57
 
@@ -174,6 +161,8 @@ plane_map["F-22A"] = F_22A
 plane_map["MB-339PAN"] = MB_339PAN
 plane_map["Su-57"] = Su_57
 plane_map["Hercules"] = Hercules
+plane_map["JAS39Gripen"] = JAS39Gripen
+plane_map["JAS39Gripen_AG"] = JAS39Gripen_AG
 
 vehicle_map["FieldHL"] = frenchpack._FIELD_HIDE
 vehicle_map["HARRIERH"] = frenchpack._FIELD_HIDE_SMALL
@@ -405,6 +394,7 @@ PRICES = {
     F_22A: 40,
     Tornado_IDS: 20,
     Tornado_GR4: 20,
+    JAS39Gripen: 26,
     # bomber
     Su_17M4: 10,
     Su_25: 15,
@@ -418,6 +408,7 @@ PRICES = {
     A_10C: 22,
     A_10C_2: 24,
     S_3B: 10,
+    JAS39Gripen_AG: 26,
     # heli
     Ka_50: 13,
     SA342M: 8,
@@ -560,6 +551,12 @@ PRICES = {
     Armor.Car_Daimler_Armored: 8,
     Armor.LT_Mk_VII_Tetrarch: 8,
     Unarmed.Tractor_M4_Hi_Speed: 2,
+    Unarmed.Carrier_Sd_Kfz_7_Tractor: 1,
+    Unarmed.LUV_Kettenrad: 1,
+    Unarmed.LUV_Kubelwagen_82: 1,
+    Unarmed.Truck_Opel_Blitz: 1,
+    Unarmed.Truck_Bedford: 1,
+    Unarmed.Truck_GMC_Jimmy_6x6_Truck: 1,
     # ship
     CV_1143_5_Admiral_Kuznetsov: 100,
     CVN_74_John_C__Stennis: 100,
@@ -801,6 +798,7 @@ CARRIER_CAPABLE = [
     Su_33,
     A_4E_C,
     S_3B,
+    S_3B_Tanker,
     E_2C,
     UH_1H,
     Mi_8MT,